+++ packages/runtime-web/src/components/src/api/ui/map/map.ts
@@ -5,13 +5,16 @@ import { requestAnimationFrame, uuid } from '../../../utils'
 import addScript from '../../../utils/add-script'
 import arrConverter from '../../../utils/array-converter'
 import boolConverter from '../../../utils/bool-converter'
+import EventBus from '../../../utils/event'
 import objConverter from '../../../utils/object-converter'
+import { sleep } from '../../../utils/sleep'
 import { attributes, properties } from './property'
 import style from './style'
 import { arrow, invertedTriangle } from './svg-icon'
 
 const AMAP_SDK = 'https://webapi.amap.com/maps'
 const AMAP_VERSION = '2.0'
+const LOAD_EVENT = 'LOAD'
 
 function anchorToOffset(x, y, w, h) {
   if (x === undefined) {
@@ -51,6 +54,8 @@ export default class Map extends BaseElement {
   mapId = 'map-' + uuid()
   // 对应高德地图中 setFitView 方法中的 avoid 参数
   fitViewAvoid = [60, 60, 60, 60]
+  // 某些方法的执行依赖地图初始化完毕，在这里实例化一个事件对象，用于通知这些方法
+  private drawEvent = new EventBus()
 
   constructor() {
     super()
@@ -81,6 +86,7 @@ export default class Map extends BaseElement {
       requestAnimationFrame(() => {
         this.drawMap(res)
         this.drawed = true
+        this.drawEvent.emit(LOAD_EVENT)
       })
     })
   }
@@ -270,14 +276,16 @@ export default class Map extends BaseElement {
   __handleRegionChange() {
     const { map } = this
 
-    map.on('dragstart', () =>
+    map.on('dragstart', (e) =>
       this.emitRegionChange({
-        type: 'begin'
+        type: 'begin',
+        causedBy: 'gesture'
       })
     )
     map.on('dragend', () =>
       this.emitRegionChange({
-        type: 'end'
+        type: 'end',
+        causedBy: 'drag'
       })
     )
   }
@@ -706,25 +714,42 @@ export default class Map extends BaseElement {
 
   getCenterLocation(options) {
     const { success, complete, fail } = options
-
-    if (this.map) {
+    const getCenterInfo = () => {
       const { lat, lng } = this.map.getCenter()
       const loc = {
         longitude: lng,
         latitude: lat,
         scale: this.map.getZoom()
       }
-      success && success(loc)
+
+      return loc
+    }
+
+    if (this.map) {
+      success && success(getCenterInfo())
       complete && complete()
     } else {
-      fail && fail({ error: 'not init' })
-      complete && complete()
+      const waitForLoad = new Promise((resolve) => {
+        this.drawEvent.on(LOAD_EVENT, () => {
+          resolve(LOAD_EVENT)
+        })
+      })
+      // 如果 map 没有初始化完成，在这里监听地图完成事件
+      // 最大等待时长 1s，超出则认为失败
+      Promise.race([sleep(1000), waitForLoad]).then((value) => {
+        if (value && value === LOAD_EVENT) success && success(getCenterInfo())
+        else {
+          fail && fail({ error: 'not init' })
+        }
+        complete && complete()
+      })
     }
   }
 
   moveToLocation(options) {
     this.emitRegionChange({
-      type: 'begin'
+      type: 'begin',
+      causedBy: 'update'
     })
 
     const { longitude, latitude } = options || {}
@@ -737,7 +762,8 @@ export default class Map extends BaseElement {
 
       setTimeout(() => {
         this.emitRegionChange({
-          type: 'end'
+          type: 'end',
+          causedBy: 'update'
         })
       }, animationDuration)
     }