+++ packages/runtime-web/src/components/src/private/page-host.ts
@@ -11,7 +11,8 @@ import { rpxToRem } from '../rpx'
 import {
   converterForPx,
   defineElement,
-  getCurrentPagePath,
+  getCurrentPageParams,
+  getPageConfig,
   shouldEnableFor
 } from '../utils'
 import boolConverter from '../utils/bool-converter'
@@ -113,6 +114,22 @@ export default class PageHost extends LitElement implements IPageHost {
   private styleElement: HTMLStyleElement
   setConfig(config: IPageConfig) {
     this.config = { ...this.config, ...config }
+    // 业务在页面打开时有覆盖页面配置的场景
+    const updatePageConfig = get(window.$MOR_APP_CONFIG, 'updatePageConfig')
+    if (updatePageConfig) {
+      const { path: pagePath, options: pageOptions } = getCurrentPageParams([
+        'path',
+        'options'
+      ])
+      const resultConfig = getPageConfig(
+        updatePageConfig,
+        pagePath,
+        pageOptions
+      )
+
+      if (typeof resultConfig === 'object')
+        this.config = Object.assign(this.config, resultConfig)
+    }
     // 强制渲染
     this.requestUpdate()
     // NOTE: 页面背景色
@@ -355,10 +372,15 @@ export default class PageHost extends LitElement implements IPageHost {
     } = this.config
 
     const pageHeaderConfig = get(window.$MOR_APP_CONFIG, 'pageHeaderConfig', {})
-    const pagePath = getCurrentPagePath()
+    const { path: pagePath, options: pageOptions } = getCurrentPageParams([
+      'path',
+      'options'
+    ])
+
     const enableShowHeader = shouldEnableFor(
       pageHeaderConfig.showHeader,
-      pagePath
+      pagePath,
+      pageOptions
     )
     // enableShowHeader 有可能为 undefined，代表用户未配置或者取数据异常
     let showHeader =
@@ -374,7 +396,8 @@ export default class PageHost extends LitElement implements IPageHost {
 
       const enableShowBack = shouldEnableFor(
         pageHeaderConfig.showBack,
-        pagePath
+        pagePath,
+        pageOptions
       )
       if (typeof enableShowBack === 'boolean') showBack = enableShowBack
     } catch (e) {}