+++ packages/create-mor/CHANGELOG.md
@@ -3,6 +3,262 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.101](https://github.com/eleme/morjs/compare/v1.0.100...v1.0.101) (2024-02-23)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.100](https://github.com/eleme/morjs/compare/v1.0.99...v1.0.100) (2024-02-23)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.99](https://github.com/eleme/morjs/compare/v1.0.98...v1.0.99) (2024-02-19)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.98](https://github.com/eleme/morjs/compare/v1.0.97...v1.0.98) (2024-02-01)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.97](https://github.com/eleme/morjs/compare/v1.0.96...v1.0.97) (2024-01-30)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.96](https://github.com/eleme/morjs/compare/v1.0.95...v1.0.96) (2024-01-17)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.95](https://github.com/eleme/morjs/compare/v1.0.95-beta.1...v1.0.95) (2024-01-16)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.95-beta.1](https://github.com/eleme/morjs/compare/v1.0.95-beta.0...v1.0.95-beta.1) (2024-01-15)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.95-beta.0](https://github.com/eleme/morjs/compare/v1.0.94...v1.0.95-beta.0) (2024-01-15)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.94](https://github.com/eleme/morjs/compare/v1.0.93...v1.0.94) (2024-01-12)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.93](https://github.com/eleme/morjs/compare/v1.0.92...v1.0.93) (2023-12-27)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.92](https://github.com/eleme/morjs/compare/v1.0.91...v1.0.92) (2023-12-20)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.91](https://github.com/eleme/morjs/compare/v1.0.90...v1.0.91) (2023-12-18)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.90](https://github.com/eleme/morjs/compare/v1.0.89...v1.0.90) (2023-12-12)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.89](https://github.com/eleme/morjs/compare/v1.0.88...v1.0.89) (2023-12-05)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.88](https://github.com/eleme/morjs/compare/v1.0.87...v1.0.88) (2023-11-16)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.87](https://github.com/eleme/morjs/compare/v1.0.86...v1.0.87) (2023-11-15)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.85](https://github.com/eleme/morjs/compare/v1.0.84...v1.0.85) (2023-11-06)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.83](https://github.com/eleme/morjs/compare/v1.0.82...v1.0.83) (2023-09-14)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.82](https://github.com/eleme/morjs/compare/v1.0.81...v1.0.82) (2023-09-08)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.81](https://github.com/eleme/morjs/compare/v1.0.80...v1.0.81) (2023-09-01)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.80](https://github.com/eleme/morjs/compare/v1.0.79...v1.0.80) (2023-08-30)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.79](https://github.com/eleme/morjs/compare/v1.0.78...v1.0.79) (2023-08-28)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.78](https://github.com/eleme/morjs/compare/v1.0.77...v1.0.78) (2023-08-18)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.77](https://github.com/eleme/morjs/compare/v1.0.76...v1.0.77) (2023-08-17)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.76](https://github.com/eleme/morjs/compare/v1.0.75...v1.0.76) (2023-08-08)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.75](https://github.com/eleme/morjs/compare/v1.0.74...v1.0.75) (2023-08-03)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.74](https://github.com/eleme/morjs/compare/v1.0.73...v1.0.74) (2023-08-01)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.73](https://github.com/eleme/morjs/compare/v1.0.72...v1.0.73) (2023-08-01)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.72](https://github.com/eleme/morjs/compare/v1.0.71...v1.0.72) (2023-08-01)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.71](https://github.com/eleme/morjs/compare/v1.0.70...v1.0.71) (2023-07-26)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
+## [1.0.70](https://github.com/eleme/morjs/compare/v1.0.69...v1.0.70) (2023-07-19)
+
+**Note:** Version bump only for package create-mor
+
+
+
+
+
 ## [1.0.69](https://github.com/eleme/morjs/compare/v1.0.68...v1.0.69) (2023-07-17)
 
 **Note:** Version bump only for package create-mor